used a button_to helper for the destroy button on the edit page
destroy action added to the BlogPost Controller as well as a redirect_to helper for the user to go back to the index route after the change is made

added a turbo confirm modal whenever a user is deleting a page with a 'are you sure?'

refactoring the 6+1 routes in the routes.rb file. as these are basic CRUD endpoints, used the resources function to refactor the code. redundant to write this everytime. DRY principle here

added a before_action to the controllers for BlogPosts controller and defined the set_blog_post method, and in this way, i was able to reuse the code across the actions for '@blog_post = BlogPost.find(params[:id])'

removed the rescue error as that's optional